feat: Create smart contract Rust bindings
Create a Rust cargo project to hold the smart contracts bindings.

The bindings are auto-generated by `ethers` ABI generator, so this crate
has a `build.rs` script to generate them.
This crate also has a basic test to ensure that the generated source can
be accessed when using the crate.

This commit also has several refactoring changes.

Because the implementation of the command-line tool is split with two
crates, one for auto-generated the smart contract bindings and the
other for the tool itself, the current Cargo.toml of the repository's
root defines a Cargo workspace.

The command-line crate is at this moment a Cargo bin place holder. This
will likely have the command-line implementation on future commits.

The smart contract was inside of the `tests` directory of the single
crate that existed, however, now it's in a directory of the repository's
root. I thought that this part of the development of the tool and isn't
just for testing.
I also thought about the names of the crates and the contract and I
renamed the contract and associated files, to hopefully be clearer aside
of shorter.

This commit also adapt the `Makefile` to the new repository's structure
and add a brief documentation of it to the README.