feat: Implement high level IPFS client
Implement a high level IPFS client to conduct the operations that the
command-line tool requires.

Added a error module to homogenize all the errors that any function or
method of this crate can return to a single type and ease the error
handling and report better error messages to the command-line users.